rename: 의미 있는 패키지 이름 변경
- step* 형식 보다 미션 주제에 대한 네이밍으로 변경
