+++ src/main/kotlin/basic/Person.kt
@@ -1,3 +1,3 @@
-package step1
+package basic
 
 data class Person(val name: String, val age: Int, var nickname: String? = name)
