+++ src/test/kotlin/basic/PersonKoTest.kt
@@ -1,4 +1,4 @@
-package step1
+package basic
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.nulls.shouldBeNull
