+++ src/test/kotlin/calculator/CalculatorTest.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
