+++ src/test/kotlin/calculator/ParserTest.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
